app/requirements.txt
emoji==2.8.0
matplotlib==3.8.0
nltk==3.8.1
numpy==1.24.3
pandas==2.1.1
parallelbar==1.3.1
pycld2==0.41
pyenchant==3.2.2
scikit_learn==1.3.0
spacy==3.6.1
textblob==0.17.1
transformers==4.33.3
torch==2.1.0
wordcloud==1.9.2
dash==2.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer
dash_bootstrap_components
dash-table==5.0.0
diskcache==5.6.3